requirements.txt
rembg
Pillow
flask[async]
onnxruntime
opencv-python
numpy
pymongo
flask-cors
bcrypt
python-dotenv
flask-jwt-extended
asyncio